requirements.txt
fastapi 
uvicorn 
sqlalchemy 
mysqlclient 
pydantic 
email-validator
python-dotenv
mysql-connector-python
jinja2
celery
mysqlclient
django
redis
requests
flower
